+++ app/controllers/api/v1/users_controller.rb
@@ -1,0 +1,48 @@
+module Api
+  module V1
+    class UsersController < ApplicationController
+      before_action :set_user, only: [:show, :update, :destroy]
+
+      def index
+        @users = User.search_user(params[:user_name]).order_user_by_name
+
+        render json: @users
+      end
+
+      def show
+        render json: @user
+      end
+
+      def create
+        @user = User.new(user_params)
+
+        if @user.save
+          render json: @user, status: :created, location: api_user_url(@user)
+        else
+          render json: @user.errors, status: :unprocessable_entity
+        end
+      end
+
+      def update
+        if @user.update(user_params)
+          render json: @user
+        else
+          render json: @user.errors, status: :unprocessable_entity
+        end
+      end
+
+      def destroy
+        @user.destroy
+      end
+
+      private
+        def set_user
+          @user = User.find(params[:id])
+        end
+
+        def user_params
+          params.require(:user).permit(:name, :email, :password, :password_confirmation)
+        end
+    end
+  end
+end
