+++ config/routes.rb
@@ -1,6 +1,11 @@
+require 'api_constraints'
+
 Rails.application.routes.draw do
   namespace :api, defaults: { format: :json } do
-    resources :stores
-    resources :users
+    scope module: :v1,
+    constraints: ApiConstraints.new(version: 1, default: true) do
+      resources :stores
+      resources :users
+    end
   end
 end
